+++ src/platform/packages/shared/kbn-visualization-utils/index.ts
@@ -14,3 +14,4 @@ export { useDebouncedValue } from './src/debounced_value';
 export { ChartType } from './src/types';
 export { getDatasourceId } from './src/get_datasource_id';
 export { mapVisToChartType } from './src/map_vis_to_chart_type';
+export { computeInterval } from './src/compute_interval';
