+++ src/platform/packages/shared/kbn-visualization-utils/tsconfig.json
@@ -13,5 +13,6 @@
     "@kbn/interpreter",
     "@kbn/data-views-plugin",
     "@kbn/es-query",
+    "@kbn/data-plugin",
   ]
 }
