+++ x-pack/platform/plugins/shared/streams_app/public/components/data_management/stream_detail_lifecycle/helpers/format_bytes.ts
@@ -6,5 +6,20 @@
  */
 
 import { formatNumber } from '@elastic/eui';
+import { i18n } from '@kbn/i18n';
 
 export const formatBytes = (value: number) => formatNumber(value, '0.0 b');
+
+export const formatIngestionRate = (bytesPerDay: number, perDayOnly = false) => {
+  const perDay = formatBytes(bytesPerDay);
+  const perMonth = formatBytes(bytesPerDay * 30);
+  if (perDayOnly)
+    return i18n.translate('xpack.streams.streamDetailOverview.ingestionRatePerDay', {
+      defaultMessage: '{perDay} / Day',
+      values: { perDay },
+    });
+  return i18n.translate('xpack.streams.streamDetailOverview.ingestionRatePerDayPerMonth', {
+    defaultMessage: '{perDay} / Day ({perMonth} / Month)',
+    values: { perDay, perMonth },
+  });
+};
