+++ x-pack/platform/plugins/shared/streams_app/public/components/entity_detail_view/index.tsx
@@ -30,6 +30,7 @@ export interface EntityViewTab {
   name: string;
   label: string;
   content: React.ReactElement;
+  background: boolean;
 }
 
 export function EntityDetailViewWithoutParams({
@@ -75,6 +76,7 @@ export function EntityDetailViewWithoutParams({
           }),
           label: tab.label,
           content: tab.content,
+          background: tab.background,
         },
       ];
     })
@@ -126,7 +128,9 @@ export function EntityDetailViewWithoutParams({
           />
         </StreamsAppPageHeader>
       </EuiFlexItem>
-      <StreamsAppPageBody>{selectedTabObject.content}</StreamsAppPageBody>
+      <StreamsAppPageBody background={selectedTabObject.background}>
+        {selectedTabObject.content}
+      </StreamsAppPageBody>
     </EuiFlexGroup>
   );
 }
