+++ x-pack/platform/plugins/shared/streams_app/public/components/esql_chart/controlled_esql_chart.tsx
@@ -62,12 +62,14 @@ export function ControlledEsqlChart<T extends string>({
   metricNames,
   chartType = 'line',
   height,
+  timerange,
 }: {
   id: string;
   result: AbortableAsyncState<UnparsedEsqlResponse>;
   metricNames: T[];
   chartType?: 'area' | 'bar' | 'line';
-  height: number;
+  height?: number;
+  timerange?: { start: number; end: number };
 }) {
   const {
     core: { uiSettings },
@@ -84,12 +86,14 @@ export function ControlledEsqlChart<T extends string>({
     [result, ...metricNames]
   );
 
+  const effectiveHeight = height ? `${height}px` : '100%';
+
   if (result.loading && !result.value?.values.length) {
     return (
       <LoadingPanel
         loading
         className={css`
-          height: ${height}px;
+          height: ${effectiveHeight};
         `}
       />
     );
@@ -97,8 +101,9 @@ export function ControlledEsqlChart<T extends string>({
 
   const xValues = allTimeseries.flatMap(({ data }) => data.map(({ x }) => x));
 
-  const min = Math.min(...xValues);
-  const max = Math.max(...xValues);
+  // todo - pull in time range here
+  const min = timerange?.start ?? Math.min(...xValues);
+  const max = timerange?.end ?? Math.max(...xValues);
 
   const isEmpty = min === 0 && max === 0;
 
@@ -115,7 +120,7 @@ export function ControlledEsqlChart<T extends string>({
     <Chart
       id={id}
       className={css`
-        height: ${height}px;
+        height: ${effectiveHeight};
       `}
     >
       <Tooltip
@@ -146,7 +151,7 @@ export function ControlledEsqlChart<T extends string>({
         }}
       />
       <Settings
-        showLegend
+        showLegend={false}
         legendPosition={Position.Bottom}
         xDomain={xDomain}
         locale={i18n.getLocale()}
@@ -173,6 +178,7 @@ export function ControlledEsqlChart<T extends string>({
           <Series
             timeZone={timeZone}
             key={serie.id}
+            color="#61A2FF"
             id={serie.id}
             xScaleType={ScaleType.Time}
             yScaleType={ScaleType.Linear}
