+++ x-pack/platform/plugins/shared/streams_app/public/components/stream_detail_view/index.tsx
@@ -45,6 +45,7 @@ export function StreamDetailViewContent({ name, tab }: { name: string; tab: stri
       label: i18n.translate('xpack.streams.streamDetailView.overviewTab', {
         defaultMessage: 'Overview',
       }),
+      background: false,
     },
     {
       name: 'dashboards',
@@ -52,6 +53,7 @@ export function StreamDetailViewContent({ name, tab }: { name: string; tab: stri
       label: i18n.translate('xpack.streams.streamDetailView.dashboardsTab', {
         defaultMessage: 'Dashboards',
       }),
+      background: true,
     },
     {
       name: 'management',
@@ -59,6 +61,7 @@ export function StreamDetailViewContent({ name, tab }: { name: string; tab: stri
       label: i18n.translate('xpack.streams.streamDetailView.managementTab', {
         defaultMessage: 'Management',
       }),
+      background: true,
     },
   ];
