+++ x-pack/platform/plugins/shared/streams_app/public/components/stream_list_view/index.tsx
@@ -47,7 +47,7 @@ export function StreamListView() {
           }
         />
       </EuiFlexItem>
-      <StreamsAppPageBody>
+      <StreamsAppPageBody background>
         <EuiFlexGroup direction="column">
           <EuiFlexItem grow={false}>
             <EuiSearchBar
