+++ x-pack/platform/plugins/shared/streams_app/tsconfig.json
@@ -27,13 +27,11 @@
     "@kbn/react-kibana-context-render",
     "@kbn/code-editor",
     "@kbn/ui-theme",
-    "@kbn/calculate-auto",
     "@kbn/kibana-react-plugin",
     "@kbn/es-query",
     "@kbn/server-route-repository-client",
     "@kbn/logging",
     "@kbn/config-schema",
-    "@kbn/calculate-auto",
     "@kbn/streams-plugin",
     "@kbn/share-plugin",
     "@kbn/code-editor",
@@ -59,6 +57,10 @@
     "@kbn/licensing-plugin",
     "@kbn/datemath",
     "@kbn/xstate-utils",
+    "@kbn/visualization-utils",
+    "@kbn/utility-types",
+    "@kbn/discover-utils",
+    "@kbn/discover-shared-plugin",
     "@kbn/core-analytics-browser",
   ]
 }
